refactor(models): centralize router Pydantic models in models.py (INV-14, #654)
Move 97 request/response BaseModel classes from 32 routers into
src/backend/models.py (Architectural Invariant #14), plus the fan_out /
loops / webhooks module-level constants their validators reference. Each
class moved verbatim — validators, default_factory, and Config preserved.

Behaviour preserved, proven two ways the move could silently break it:
- Normalized app.openapi() snapshot diff: 330 paths / 198 schemas,
  byte-identical origin/dev vs branch (schema-level proof).
- Validator-preservation unit tests for the 5 behaviour-bearing files
  (fan_out, loops, canary, audit_log, schedules) — the field_validator
  logic, default_factory dicts, and Config.from_attributes an OpenAPI
  diff can't see.

One documented exception, allowlisted in the new static guard:
canary.py::RunCycleRequest evaluates INVARIANTS (the canary library) in a
Field(description=...) at class-definition time, and the canary library
imports TaskExecutionStatus back from models — relocating it would force
models.py to `from canary import ...`, inverting the dependency direction
of a module meant to be a low-level leaf everything imports from.

Flat models.py (not a package): a models/ dir would break
tests/conftest.py's by-file model preload and the Dockerfile's by-file
COPY *.py (the #1033 crash-loop class). Scope is router models only —
db_models.py and adapters/base.py are intentional separate homes.

New / changed:
- tests/unit/test_models_centralized.py: static guard (no BaseModel under
  routers/ except the documented allowlist) + planted-violation tests
- tests/unit/test_router_model_validators.py: validator / default_factory
  / Config preservation negatives
- tests/unit/test_telegram_webhook_backfill.py: load the real models
  module in the standalone settings.py stub (was already red on dev) and
  stub the #1197 services.agent_service.capabilities import -> green
- docs/memory/architecture.md: Invariant #14 scope note + re-measured
  router/service/agents counts
