+++ src/backend/models.py
@@ -1,8 +1,10 @@
 """
 Pydantic models for the Trinity backend API.
 """
-from pydantic import BaseModel, Field
-from typing import Dict, List, Optional
+import re
+
+from pydantic import BaseModel, EmailStr, Field, field_validator
+from typing import Dict, List, Literal, Optional
 from datetime import datetime
 from enum import Enum
 
@@ -874,3 +876,1161 @@ class CompatibilityFixResponse(BaseModel):
     fixed: bool
     message: str
     uncommitted: bool = True
+
+
+# =============================================================================
+# Router-relocated request/response models (#654, INV-14)
+# Each section below was moved verbatim from its router so Pydantic models
+# live in one place (Architectural Invariant #14). One exception remains in
+# routers/canary.py (RunCycleRequest) — see test_models_centralized.py.
+# =============================================================================
+
+
+# =============================================================================
+# Agent Files Models (routers/agent_files.py)
+# =============================================================================
+
+
+class FileUpdateRequest(BaseModel):
+    """Request body for file updates."""
+    content: str
+
+
+class CreateFolderRequest(BaseModel):
+    """Request body for folder creation."""
+    path: str
+
+
+# =============================================================================
+# Agent Rename Models (routers/agent_rename.py)
+# =============================================================================
+
+
+class RenameAgentRequest(BaseModel):
+    """Request body for agent rename."""
+    new_name: str
+
+
+# =============================================================================
+# Agent Ssh Models (routers/agent_ssh.py)
+# =============================================================================
+
+
+class SshAccessRequest(BaseModel):
+    """Request body for SSH access."""
+    ttl_hours: float = 4.0
+    auth_method: str = "key"  # "key" for SSH key, "password" for ephemeral password
+    public_key: Optional[str] = None  # Required for key auth — client-supplied OpenSSH public key
+
+
+# =============================================================================
+# Agents Models (routers/agents.py)
+# =============================================================================
+
+
+class HeartbeatPayload(BaseModel):
+    """Lightweight liveness payload POSTed by the agent every ~5s."""
+    memory_mb: Optional[float] = None
+    active_executions: Optional[int] = None
+    uptime_s: Optional[float] = None
+
+
+# =============================================================================
+# Audit Log Models (routers/audit_log.py)
+# =============================================================================
+
+
+class AuditLogEntry(BaseModel):
+    """Single audit log row as returned to API clients."""
+
+    id: int
+    event_id: str
+    event_type: str
+    event_action: str
+    actor_type: str
+    actor_id: Optional[str] = None
+    actor_email: Optional[str] = None
+    actor_ip: Optional[str] = None
+    mcp_key_id: Optional[str] = None
+    mcp_key_name: Optional[str] = None
+    mcp_scope: Optional[str] = None
+    target_type: Optional[str] = None
+    target_id: Optional[str] = None
+    timestamp: str
+    details: Optional[dict] = None
+    request_id: Optional[str] = None
+    source: str
+    endpoint: Optional[str] = None
+    previous_hash: Optional[str] = None
+    entry_hash: Optional[str] = None
+    created_at: Optional[str] = None
+
+
+class AuditLogListResponse(BaseModel):
+    """Paginated list response."""
+
+    entries: List[AuditLogEntry]
+    total: int
+    limit: int
+    offset: int
+
+
+class AuditLogStatsResponse(BaseModel):
+    """Aggregate counts."""
+
+    total: int
+    by_event_type: dict = Field(default_factory=dict)
+    by_actor_type: dict = Field(default_factory=dict)
+
+
+class AuditHeatmapCell(BaseModel):
+    """Single populated bucket in the 7×24 dow×hour heatmap."""
+
+    dow: int = Field(..., ge=0, le=6, description="Weekday (0=Sunday)")
+    hour: int = Field(..., ge=0, le=23, description="Hour 0–23 UTC")
+    count: int = Field(..., ge=0)
+
+
+class AuditHeatmapResponse(BaseModel):
+    """Sparse 7×24 dow×hour heatmap. Zero-count cells omitted."""
+
+    cells: List[AuditHeatmapCell]
+    total: int
+    max_count: int
+
+
+class AuditCalendarDay(BaseModel):
+    """Single populated day in the calendar heatmap."""
+
+    date: str = Field(..., description="UTC date, ISO 'YYYY-MM-DD'")
+    count: int = Field(..., ge=0)
+
+
+class AuditCalendarResponse(BaseModel):
+    """Sparse per-day calendar heatmap (GitHub-style). Quiet days omitted."""
+
+    days: List[AuditCalendarDay]
+    total: int
+    max_count: int
+
+
+class AuditVerifyResponse(BaseModel):
+    """Hash chain verification result."""
+
+    valid: bool
+    checked: int
+    first_invalid_id: Optional[int] = None
+
+
+# =============================================================================
+# Avatar Models (routers/avatar.py)
+# =============================================================================
+
+
+class AvatarGenerateRequest(BaseModel):
+    identity_prompt: str
+
+
+# =============================================================================
+# Canary Models (routers/canary.py)
+# =============================================================================
+
+
+class CanaryViolation(BaseModel):
+    """Single canary_violations row as returned to API clients."""
+
+    id: int
+    invariant_id: str
+    tier: str
+    severity: str
+    snapshot_time: str
+    observed_state: dict = Field(default_factory=dict)
+    signal_query: Optional[str] = None
+    created_at: Optional[str] = None
+
+
+class CanaryViolationListResponse(BaseModel):
+    """Paginated list response."""
+
+    violations: List[CanaryViolation]
+    total: int
+    limit: int
+    offset: int
+
+
+class CanaryStatsResponse(BaseModel):
+    """Aggregate violation counts for dashboard tiles."""
+
+    total: int
+    by_invariant: dict = Field(default_factory=dict)
+    by_severity: dict = Field(default_factory=dict)
+
+
+class CycleViolation(BaseModel):
+    """One violation persisted during a run-cycle call."""
+
+    id: int
+    invariant_id: str
+    tier: str
+    severity: str
+    snapshot_time: str
+    observed_state: dict
+    signal_query: Optional[str] = None
+
+
+class CycleTransition(BaseModel):
+    """A green→red transition detected this cycle.
+
+    `CanaryService` posts exactly one Slack webhook message per entry,
+    mapping severity to the message styling. Surfaced here so the run-cycle
+    response mirrors what the service actually emitted.
+    """
+
+    invariant_id: str
+    severity: str
+    violations_in_cycle: int
+    previous_violation_at: Optional[str] = Field(
+        None,
+        description=(
+            "snapshot_time of the most recent prior violation for this "
+            "invariant; null if the invariant has never violated before."
+        ),
+    )
+
+
+class RunCycleResponse(BaseModel):
+    """Result of one canary cycle."""
+
+    snapshot_time: str
+    cycle_duration_ms: int
+    # Invariants this cycle attempted (= the request's `invariants` filter,
+    # or all registered ids if unfiltered). Whether each one *fired* is
+    # surfaced via `violations` and `transitions`. Sources that were down
+    # this cycle are listed in `sources_unavailable` — invariants that
+    # depend on them returned no violations regardless of state.
+    checks_run: List[str]
+    sources_unavailable: List[str]
+    violations: List[CycleViolation]
+    transitions: List[CycleTransition]
+
+
+# =============================================================================
+# Event Subscriptions Models (routers/event_subscriptions.py)
+# =============================================================================
+
+
+class EmitEventRequest(BaseModel):
+    """Request body for emitting an event."""
+    event_type: str  # Namespaced event type (e.g., "prediction.resolved")
+    payload: Optional[dict] = None  # Structured data
+
+
+# =============================================================================
+# Fan Out Models (routers/fan_out.py)
+# =============================================================================
+
+
+TASK_ID_RE = re.compile(r"^[a-zA-Z0-9_-]{1,64}$")
+
+
+MAX_TASKS = 50
+
+
+MAX_CONCURRENCY = 10
+
+
+class FanOutTask(BaseModel):
+    """A single task in a fan-out request."""
+    id: str
+    message: str = Field(..., min_length=1, max_length=100_000)
+
+    @field_validator("id")
+    @classmethod
+    def validate_task_id(cls, v: str) -> str:
+        if not TASK_ID_RE.match(v):
+            raise ValueError(
+                f"Task ID must be 1-64 alphanumeric characters, hyphens, or underscores: '{v}'"
+            )
+        return v
+
+
+class FanOutRequest(BaseModel):
+    """Request model for fan-out parallel task execution."""
+    tasks: List[FanOutTask]
+    agent: str = "self"
+    # Optional overall fan-out deadline. When None, no outer deadline is
+    # applied — each sub-task is still bounded by the target agent's
+    # configured execution_timeout_seconds (TIMEOUT-001).
+    timeout_seconds: Optional[int] = None
+    max_concurrency: int = 3
+    policy: str = "best-effort"
+    model: Optional[str] = None
+    system_prompt: Optional[str] = None
+    allowed_tools: Optional[List[str]] = None
+
+    @field_validator("tasks")
+    @classmethod
+    def validate_tasks(cls, v: List[FanOutTask]) -> List[FanOutTask]:
+        if len(v) == 0:
+            raise ValueError("At least one task is required")
+        if len(v) > MAX_TASKS:
+            raise ValueError(f"Maximum {MAX_TASKS} tasks per fan-out")
+        # Check for duplicate IDs
+        ids = [t.id for t in v]
+        if len(ids) != len(set(ids)):
+            dupes = [i for i in ids if ids.count(i) > 1]
+            raise ValueError(f"Duplicate task IDs: {set(dupes)}")
+        return v
+
+    @field_validator("max_concurrency")
+    @classmethod
+    def validate_concurrency(cls, v: int) -> int:
+        if v < 1 or v > MAX_CONCURRENCY:
+            raise ValueError(f"max_concurrency must be between 1 and {MAX_CONCURRENCY}")
+        return v
+
+    @field_validator("timeout_seconds")
+    @classmethod
+    def validate_timeout(cls, v: Optional[int]) -> Optional[int]:
+        if v is None:
+            return v
+        if v < 10 or v > 3600:
+            raise ValueError("timeout_seconds must be between 10 and 3600")
+        return v
+
+    @field_validator("policy")
+    @classmethod
+    def validate_policy(cls, v: str) -> str:
+        if v != "best-effort":
+            raise ValueError("Only 'best-effort' policy is supported")
+        return v
+
+
+class FanOutTaskResponse(BaseModel):
+    """Result of a single fan-out subtask."""
+    id: str
+    status: str
+    response: Optional[str] = None
+    error: Optional[str] = None
+    error_code: Optional[str] = None
+    execution_id: Optional[str] = None
+    cost: Optional[float] = None
+    context_used: Optional[int] = None
+    duration_ms: Optional[int] = None
+
+
+class FanOutResponse(BaseModel):
+    """Aggregated fan-out result."""
+    fan_out_id: str
+    status: str
+    total: int
+    completed: int
+    failed: int
+    results: List[FanOutTaskResponse]
+
+
+# =============================================================================
+# Git Models (routers/git.py)
+# =============================================================================
+
+
+class GitSyncRequest(BaseModel):
+    """Request body for git sync operation."""
+    message: Optional[str] = None  # Custom commit message
+    paths: Optional[List[str]] = None  # Specific paths to sync
+    strategy: Optional[str] = "normal"  # "normal", "pull_first", "force_push"
+
+
+class GitPullRequest(BaseModel):
+    """Request body for git pull operation."""
+    strategy: Optional[str] = "clean"  # "clean", "stash_reapply", "force_reset"
+
+
+class GitInitializeRequest(BaseModel):
+    """Request body for git initialization."""
+    repo_owner: str  # GitHub username or organization
+    repo_name: str  # Repository name
+    create_repo: bool = True  # Whether to create the repository if it doesn't exist
+    private: bool = True  # Whether the new repository should be private
+    description: Optional[str] = None  # Repository description
+
+
+class GitHubPATRequest(BaseModel):
+    """Request body for setting agent GitHub PAT."""
+    pat: str
+
+
+class AutoSyncToggle(BaseModel):
+    enabled: bool
+
+
+class FreezeSchedulesToggle(BaseModel):
+    enabled: bool
+
+
+# =============================================================================
+# Image Generation Models (routers/image_generation.py)
+# =============================================================================
+
+
+class ImageGenerateRequest(BaseModel):
+    prompt: str
+    use_case: Optional[str] = "general"
+    aspect_ratio: Optional[str] = "1:1"
+    refine_prompt: Optional[bool] = True
+
+
+# =============================================================================
+# Internal Models (routers/internal.py)
+# =============================================================================
+
+
+class ActivityTrackRequest(BaseModel):
+    """Request model for tracking activity start."""
+    agent_name: str
+    activity_type: str  # e.g., "schedule_start"
+    user_id: Optional[int] = None
+    triggered_by: str = "schedule"  # schedule, manual, user, agent, system
+    related_execution_id: Optional[str] = None
+    details: Optional[Dict] = None
+
+
+class ActivityCompleteRequest(BaseModel):
+    """Request model for completing an activity."""
+    status: str = ActivityState.COMPLETED  # ActivityState: completed, failed
+    details: Optional[Dict] = None
+    error: Optional[str] = None
+
+
+class InternalTaskExecutionRequest(BaseModel):
+    """Request model for internal task execution via TaskExecutionService."""
+    agent_name: str
+    message: str
+    triggered_by: str = "schedule"
+    model: Optional[str] = None
+    timeout_seconds: Optional[int] = None  # TIMEOUT-001: None = use agent's config (default 15 min)
+    allowed_tools: Optional[List[str]] = None
+    execution_id: Optional[str] = None
+    async_mode: bool = False
+    # #171: optional schedule metadata surfaced in the agent's execution context block.
+    schedule_name: Optional[str] = None
+    schedule_cron: Optional[str] = None
+    schedule_next_run: Optional[str] = None
+    attempt: Optional[int] = None
+
+
+class ValidateExecutionRequest(BaseModel):
+    """Request model for triggering execution validation."""
+    execution_id: str
+    agent_name: str
+    schedule_id: str
+    original_message: str
+    execution_response: str
+    custom_prompt: Optional[str] = None
+    timeout_seconds: int = 120
+
+
+class InternalAuditRequest(BaseModel):
+    """Request model for audit log entries from MCP server."""
+    event_type: str          # AuditEventType value
+    event_action: str        # e.g. "tool_call"
+    source: str = "mcp"      # Always "mcp" for MCP server calls
+    # MCP auth context
+    mcp_key_id: Optional[str] = None
+    mcp_key_name: Optional[str] = None
+    mcp_scope: Optional[str] = None
+    actor_agent_name: Optional[str] = None
+    # Target
+    target_type: Optional[str] = None
+    target_id: Optional[str] = None
+    # Details
+    details: Optional[Dict] = None
+
+
+# =============================================================================
+# Logs Models (routers/logs.py)
+# =============================================================================
+
+
+class RetentionConfig(BaseModel):
+    """Retention configuration."""
+    retention_days: int = Field(..., ge=1, le=3650, description="Days to retain logs")
+    archive_enabled: bool = Field(..., description="Whether archival is enabled")
+    cleanup_hour: int = Field(..., ge=0, le=23, description="Hour (UTC) to run nightly archival")
+
+
+class ArchiveRequest(BaseModel):
+    """Manual archive request."""
+    retention_days: Optional[int] = Field(None, ge=1, le=3650, description="Override retention days")
+    delete_after_archive: bool = Field(True, description="Delete originals after archiving")
+
+
+# =============================================================================
+# Loops Models (routers/loops.py)
+# =============================================================================
+
+
+MAX_RUNS_LIMIT = 100
+
+
+MAX_MESSAGE_LEN = 100_000
+
+
+MAX_DELAY_SECONDS = 3600
+
+
+MAX_TIMEOUT_PER_RUN = 7200
+
+
+MAX_STOP_SIGNAL_LEN = 200
+
+
+class StartLoopRequest(BaseModel):
+    message: str = Field(..., min_length=1, max_length=MAX_MESSAGE_LEN)
+    max_runs: int = Field(..., ge=1, le=MAX_RUNS_LIMIT)
+    stop_signal: Optional[str] = Field(default=None, max_length=MAX_STOP_SIGNAL_LEN)
+    delay_seconds: int = Field(default=0, ge=0, le=MAX_DELAY_SECONDS)
+    timeout_per_run: Optional[int] = Field(default=None, ge=10, le=MAX_TIMEOUT_PER_RUN)
+    model: Optional[str] = None
+    allowed_tools: Optional[List[str]] = None
+
+    @field_validator("stop_signal")
+    @classmethod
+    def _normalize_stop_signal(cls, v: Optional[str]) -> Optional[str]:
+        if v is None:
+            return None
+        v = v.strip()
+        return v or None  # empty after strip → fixed mode
+
+
+class StartLoopResponse(BaseModel):
+    loop_id: str
+    status: str
+    agent_name: str
+    max_runs: int
+
+
+class LoopRunResponse(BaseModel):
+    run_number: int
+    execution_id: Optional[str] = None
+    status: str
+    response_preview: Optional[str] = None
+    cost: Optional[float] = None
+    duration_ms: Optional[int] = None
+    error: Optional[str] = None
+    started_at: str
+    completed_at: Optional[str] = None
+
+
+class LoopStatusResponse(BaseModel):
+    loop_id: str
+    agent_name: str
+    status: str
+    max_runs: int
+    runs_completed: int
+    stop_reason: Optional[str] = None
+    last_response: Optional[str] = None
+    error: Optional[str] = None
+    runs: List[LoopRunResponse]
+    created_at: str
+    started_at: Optional[str] = None
+    completed_at: Optional[str] = None
+
+
+class StopLoopResponse(BaseModel):
+    loop_id: str
+    status: str  # "stopping" | "already_done"
+
+
+# =============================================================================
+# Messages Models (routers/messages.py)
+# =============================================================================
+
+
+class SendMessageRequest(BaseModel):
+    """Request to send a proactive message to a user."""
+    recipient_email: EmailStr = Field(
+        ...,
+        description="Verified email of the recipient. Must be in agent_sharing with allow_proactive=1."
+    )
+    text: str = Field(
+        ...,
+        min_length=1,
+        max_length=4096,
+        description="Message content (max 4096 characters)"
+    )
+    channel: Literal["auto", "telegram", "slack", "web"] = Field(
+        default="auto",
+        description="Target channel. 'auto' tries channels in order: telegram -> slack -> web"
+    )
+    reply_to_thread: bool = Field(
+        default=False,
+        description="Continue in last thread if one exists (channel-dependent)"
+    )
+
+
+class SendMessageResponse(BaseModel):
+    """Response from sending a proactive message."""
+    success: bool
+    channel: str
+    message_id: Optional[str] = None
+    error: Optional[str] = None
+
+
+class ProactiveShareUpdate(BaseModel):
+    """Request to update allow_proactive flag for a share."""
+    email: EmailStr
+    allow_proactive: bool
+
+
+class ProactiveSharesResponse(BaseModel):
+    """List of emails with proactive messaging enabled."""
+    agent_name: str
+    emails: list[str]
+
+
+# =============================================================================
+# Notifications Models (routers/notifications.py)
+# =============================================================================
+
+
+class DismissAllRequest(BaseModel):
+    """Body for bulk-dismissing notifications (#1017)."""
+    agent_name: Optional[str] = None
+
+
+# =============================================================================
+# Operator Queue Models (routers/operator_queue.py)
+# =============================================================================
+
+
+class OperatorResponse(BaseModel):
+    """Body for responding to a queue item."""
+    response: str
+    response_text: Optional[str] = None
+
+
+class BulkCancelRequest(BaseModel):
+    """Body for bulk-cancelling pending queue items (#1017).
+
+    The client sends the ids it actually rendered, so a sync-loop race can
+    never cancel items the operator never saw.
+    """
+    ids: List[str] = Field(..., min_length=1, max_length=500)
+
+
+class ClearResolvedRequest(BaseModel):
+    """Body for clearing the Resolved tab (#1017)."""
+    agent_name: Optional[str] = None
+
+
+# =============================================================================
+# Paid Models (routers/paid.py)
+# =============================================================================
+
+
+class PaidChatRequest(BaseModel):
+    message: str
+    session_id: Optional[str] = None
+
+
+# =============================================================================
+# Public Models (routers/public.py)
+# =============================================================================
+
+
+class PublicChatHistoryResponse(BaseModel):
+    """Response model for chat history endpoint."""
+    messages: List[dict]
+    session_id: str
+    message_count: int
+
+
+class ClearSessionResponse(BaseModel):
+    """Response model for clear session endpoint."""
+    cleared: bool
+    new_session_id: Optional[str] = None
+
+
+# =============================================================================
+# Public Memory Models (routers/public_memory.py)
+# =============================================================================
+
+
+class WriteUserMemoryRequest(BaseModel):
+    execution_id: str = Field(..., min_length=1, max_length=200)
+    memory_text: str = Field(..., max_length=8000)
+
+
+# =============================================================================
+# Schedules Models (routers/schedules.py)
+# =============================================================================
+
+
+class ScheduleUpdateRequest(BaseModel):
+    """Request model for updating a schedule."""
+    name: Optional[str] = None
+    cron_expression: Optional[str] = None
+    message: Optional[str] = None
+    enabled: Optional[bool] = None
+    timezone: Optional[str] = None
+    description: Optional[str] = None
+    timeout_seconds: Optional[int] = None
+    allowed_tools: Optional[List[str]] = None
+    model: Optional[str] = None  # Model override (MODEL-001)
+    # Retry configuration (RETRY-001)
+    max_retries: Optional[int] = None
+    retry_delay_seconds: Optional[int] = None
+    # Validation configuration (VALIDATE-001)
+    validation_enabled: Optional[bool] = None
+    validation_prompt: Optional[str] = None
+    validation_timeout_seconds: Optional[int] = None
+
+
+class ScheduleResponse(BaseModel):
+    """Response model for schedule data."""
+    id: str
+    agent_name: str
+    name: str
+    cron_expression: str
+    message: str
+    enabled: bool
+    timezone: str
+    description: Optional[str]
+    created_at: datetime
+    updated_at: datetime
+    last_run_at: Optional[datetime]
+    next_run_at: Optional[datetime]
+    # #913: null means "inherit from agent_ownership.execution_timeout_seconds".
+    timeout_seconds: Optional[int] = None
+    allowed_tools: Optional[List[str]] = None
+    model: Optional[str] = None  # Model override (MODEL-001)
+    # Validation configuration (VALIDATE-001)
+    validation_enabled: bool = False
+    validation_prompt: Optional[str] = None
+    validation_timeout_seconds: int = 120
+
+    class Config:
+        from_attributes = True
+
+
+class ExecutionSummary(BaseModel):
+    """Lightweight execution response for list views - excludes large text fields.
+
+    Used by GET /api/agents/{name}/executions for fast list loading.
+    Full details available via GET /api/agents/{name}/executions/{id}.
+    """
+    id: str
+    schedule_id: str
+    agent_name: str
+    status: str
+    started_at: datetime
+    completed_at: Optional[datetime]
+    duration_ms: Optional[int]
+    message: str
+    triggered_by: str
+    # Observability fields (small)
+    context_used: Optional[int] = None
+    context_max: Optional[int] = None
+    cost: Optional[float] = None
+    # Origin tracking (small) - AUDIT-001
+    source_user_id: Optional[int] = None
+    source_user_email: Optional[str] = None
+    source_agent_name: Optional[str] = None
+    source_mcp_key_id: Optional[str] = None
+    source_mcp_key_name: Optional[str] = None
+    # Session resume (small) - EXEC-023
+    claude_session_id: Optional[str] = None
+    # Model selection (small) - MODEL-001
+    model_used: Optional[str] = None
+    # Fan-out linkage (small) - FANOUT-001
+    fan_out_id: Optional[str] = None
+    # Validation tracking (small) - VALIDATE-001
+    business_status: Optional[str] = None  # pending_validation, validated, failed_validation, skipped
+    validation_execution_id: Optional[str] = None
+    # Auto-compact observability (Bundle B) - small JSON list
+    compact_metadata: Optional[str] = None
+
+    # EXCLUDED (large fields - fetch via /executions/{id}):
+    # - response: Optional[str]      # Full response text
+    # - error: Optional[str]         # Full error text
+    # - tool_calls: Optional[str]    # JSON array of tool calls
+    # - execution_log: Optional[str] # Full Claude Code transcript
+
+    class Config:
+        from_attributes = True
+
+
+class ExecutionResponse(BaseModel):
+    """Full response model for execution data - includes all fields.
+
+    Used by GET /api/agents/{name}/executions/{id} for single execution details.
+    """
+    id: str
+    schedule_id: str
+    agent_name: str
+    status: str
+    started_at: datetime
+    completed_at: Optional[datetime]
+    duration_ms: Optional[int]
+    message: str
+    response: Optional[str]
+    error: Optional[str]
+    triggered_by: str
+    # Observability fields
+    context_used: Optional[int] = None
+    context_max: Optional[int] = None
+    cost: Optional[float] = None
+    tool_calls: Optional[str] = None
+    execution_log: Optional[str] = None  # Full Claude Code execution transcript (JSON)
+    # Origin tracking - AUDIT-001
+    source_user_id: Optional[int] = None
+    source_user_email: Optional[str] = None
+    source_agent_name: Optional[str] = None
+    source_mcp_key_id: Optional[str] = None
+    source_mcp_key_name: Optional[str] = None
+    # Session resume - EXEC-023
+    claude_session_id: Optional[str] = None
+    # Model selection - MODEL-001
+    model_used: Optional[str] = None
+    # Fan-out linkage - FANOUT-001
+    fan_out_id: Optional[str] = None
+    # Validation tracking - VALIDATE-001
+    business_status: Optional[str] = None
+    validated_at: Optional[datetime] = None
+    validation_execution_id: Optional[str] = None
+    validates_execution_id: Optional[str] = None
+    # Auto-compact observability (Bundle B)
+    compact_metadata: Optional[str] = None
+
+    class Config:
+        from_attributes = True
+
+
+class WebhookStatusResponse(BaseModel):
+    """Webhook configuration for a schedule."""
+    schedule_id: str
+    has_token: bool
+    webhook_enabled: bool
+    webhook_url: Optional[str] = None
+
+
+# =============================================================================
+# Sessions Models (routers/sessions.py)
+# =============================================================================
+
+
+class CreateSessionRequest(BaseModel):
+    """Optional body for POST /session. All fields optional."""
+
+    subscription_id: Optional[str] = None
+
+
+class SessionMessageRequest(BaseModel):
+    """Body for the turn endpoint."""
+
+    message: str = Field(..., min_length=1)
+    model: Optional[str] = None
+    timeout_seconds: Optional[int] = None
+    # File attachments — same shape as ParallelTaskRequest.files (#364).
+    # Images become vision blocks for the model; non-images are written
+    # into the agent workspace and a "[File uploaded by X]: name (size)
+    # saved to path" line is appended to the prompt so the agent can
+    # `Read` them. (Phase 5.2 file-upload parity with Chat.)
+    files: Optional[list] = None
+
+
+# =============================================================================
+# Settings Models (routers/settings.py)
+# =============================================================================
+
+
+class ApiKeyUpdate(BaseModel):
+    """Request body for updating an API key."""
+    api_key: str
+
+
+class ApiKeyTest(BaseModel):
+    """Request body for testing an API key."""
+    api_key: str
+
+
+class OpsSettingsUpdate(BaseModel):
+    """Request body for updating ops settings."""
+    settings: Dict[str, str]
+
+
+class SlackSettingsUpdate(BaseModel):
+    """Request body for updating Slack settings."""
+    client_id: str = None
+    client_secret: str = None
+    signing_secret: str = None
+
+
+class SlackConnectRequest(BaseModel):
+    """Request body for connecting Slack transport."""
+    app_token: Optional[str] = None  # xapp-... for Socket Mode
+    transport_mode: Optional[str] = None  # "socket" or "webhook"
+
+
+class GitHubTemplateEntry(BaseModel):
+    """A single GitHub template entry."""
+    github_repo: str
+    display_name: str = ""
+    description: str = ""
+
+
+class GitHubTemplatesUpdate(BaseModel):
+    """Request body for updating GitHub templates."""
+    templates: List[GitHubTemplateEntry]
+
+
+class McpUrlUpdate(BaseModel):
+    """Request body for updating the MCP server URL."""
+    url: str
+
+
+class AgentQuotaUpdate(BaseModel):
+    """Request body for updating per-role agent quotas."""
+    max_agents_creator: Optional[str] = None
+    max_agents_operator: Optional[str] = None
+    max_agents_user: Optional[str] = None
+
+
+# =============================================================================
+# Setup Models (routers/setup.py)
+# =============================================================================
+
+
+class SetAdminPasswordRequest(BaseModel):
+    """Request body for creating the admin account at first-time setup.
+
+    `email` is **required** (trinity-enterprise#49): it becomes the admin's
+    sign-in identity (login with email + password instead of the fixed 'admin')
+    and is the contact used for the optional operator intake. The remaining
+    operator-profile fields (company/name/role/use_case) stay optional and are
+    only forwarded to the hosted intake endpoint when `consent_updates` is true.
+    """
+    password: str = Field(..., max_length=128)
+    confirm_password: str = Field(..., max_length=128)
+    # Required admin email — sign-in identity. Shape validated in the handler so
+    # a typo / blank value yields a clean 400 (a missing field yields a 422).
+    email: str = Field(..., max_length=254)
+    # Optional operator profile — all skippable; setup completes without them.
+    company: Optional[str] = Field(None, max_length=200)
+    name: Optional[str] = Field(None, max_length=200)
+    role: Optional[str] = Field(None, max_length=200)
+    use_case: Optional[str] = Field(None, max_length=500)
+    # Affirmative, opt-in consent to occasionally receive security & product
+    # updates. ONLY when true is anything submitted to the hosted intake.
+    consent_updates: bool = False
+
+
+# =============================================================================
+# Sharing Models (routers/sharing.py)
+# =============================================================================
+
+
+class AccessPolicy(BaseModel):
+    require_email: bool
+    open_access: bool
+    group_auth_mode: str = "none"  # 'none' or 'any_verified'
+
+
+class AccessPolicyUpdate(BaseModel):
+    require_email: bool
+    open_access: bool
+    group_auth_mode: str = "none"  # 'none' or 'any_verified'
+
+
+class AccessRequest(BaseModel):
+    id: str
+    agent_name: str
+    email: str
+    channel: str | None = None
+    requested_at: str
+    status: str
+
+
+class AccessRequestDecision(BaseModel):
+    approve: bool
+
+
+# =============================================================================
+# Slack Models (routers/slack.py)
+# =============================================================================
+
+
+class SlackEventResponse(BaseModel):
+    """Response to Slack events (always return 200)."""
+    ok: bool = True
+    challenge: Optional[str] = None
+
+
+# =============================================================================
+# Telegram Models (routers/telegram.py)
+# =============================================================================
+
+
+class TelegramWebhookResponse(BaseModel):
+    ok: bool = True
+
+
+class TelegramBindingResponse(BaseModel):
+    agent_name: str
+    bot_username: Optional[str] = None
+    bot_id: Optional[str] = None
+    webhook_url: Optional[str] = None
+    bot_link: Optional[str] = None
+    configured: bool = False
+    group_count: int = 0
+    warning: Optional[str] = None
+
+
+class TelegramConfigureRequest(BaseModel):
+    bot_token: str
+
+
+class TelegramTestRequest(BaseModel):
+    chat_id: Optional[str] = None
+    message: str = "Hello from Trinity! Your Telegram bot is configured correctly."
+
+
+class TelegramGroupConfigResponse(BaseModel):
+    id: int
+    chat_id: str
+    chat_title: Optional[str] = None
+    chat_type: str = "group"
+    trigger_mode: str = "mention"
+    welcome_enabled: bool = False
+    welcome_text: Optional[str] = None
+    is_active: bool = True
+
+
+class TelegramGroupConfigUpdateRequest(BaseModel):
+    trigger_mode: Optional[str] = None
+    welcome_enabled: Optional[bool] = None
+    welcome_text: Optional[str] = None
+
+
+class TelegramGroupMessageRequest(BaseModel):
+    """Request model for proactive group messaging (Issue #349)."""
+    message: str
+
+
+# =============================================================================
+# Users Models (routers/users.py)
+# =============================================================================
+
+
+class UserRoleUpdate(BaseModel):
+    role: str
+
+
+class UpdateMyEmailRequest(BaseModel):
+    email: str
+
+
+# =============================================================================
+# Voice Models (routers/voice.py)
+# =============================================================================
+
+
+class VoiceStartRequest(BaseModel):
+    session_id: Optional[str] = None  # Existing chat session to continue
+    voice_name: Optional[str] = None  # Gemini voice name (e.g. "Kore", "Puck")
+    workspace_mode: bool = False       # Enable canvas panel tools
+
+
+class VoiceStartResponse(BaseModel):
+    voice_session_id: str
+    websocket_url: str
+    chat_session_id: str
+
+
+class VoiceStopRequest(BaseModel):
+    voice_session_id: str
+
+
+class VoiceStopResponse(BaseModel):
+    transcript: list
+    messages_saved: int
+    duration_seconds: float
+
+
+# =============================================================================
+# Voip Models (routers/voip.py)
+# =============================================================================
+
+
+class VoipConfigureRequest(BaseModel):
+    account_sid: str
+    auth_token: str
+    from_number: str
+    daily_call_cap: Optional[int] = None
+
+
+class VoipBindingResponse(BaseModel):
+    agent_name: str
+    configured: bool
+    account_sid: Optional[str] = None
+    from_number: Optional[str] = None
+    daily_call_cap: Optional[int] = None
+    display_name: Optional[str] = None
+    enabled: Optional[bool] = None
+
+
+class VoipCallRequest(BaseModel):
+    to_number: str
+    context: Optional[str] = None
+    process_transcript: bool = True
+
+
+# =============================================================================
+# Webhooks Models (routers/webhooks.py)
+# =============================================================================
+
+
+CONTEXT_MAX_CHARS = 4000
+
+
+class WebhookTriggerRequest(BaseModel):
+    """Optional body for a webhook trigger call."""
+    context: Optional[str] = Field(
+        default=None,
+        description="Additional context appended to the schedule message.",
+        max_length=CONTEXT_MAX_CHARS,
+    )
+    metadata: Optional[dict] = Field(
+        default=None,
+        description="Arbitrary key/value metadata stored on the execution record.",
+    )
+
+
+# =============================================================================
+# Whatsapp Models (routers/whatsapp.py)
+# =============================================================================
+
+
+class WhatsAppBindingResponse(BaseModel):
+    agent_name: str
+    configured: bool = False
+    account_sid: Optional[str] = None
+    from_number: Optional[str] = None
+    messaging_service_sid: Optional[str] = None
+    display_name: Optional[str] = None
+    is_sandbox: bool = False
+    webhook_url: Optional[str] = None
+    warning: Optional[str] = None
+
+
+class WhatsAppConfigureRequest(BaseModel):
+    account_sid: str
+    auth_token: str
+    from_number: str
+    messaging_service_sid: Optional[str] = None
+
+
+class WhatsAppTestRequest(BaseModel):
+    to_number: Optional[str] = None
+    message: str = "Hello from Trinity! Your WhatsApp integration is configured correctly."
