+++ src/backend/routers/fan_out.py
@@ -7,15 +7,13 @@ POST /api/agents/{name}/fan-out
 """
 
 import logging
-import re
-from typing import List, Optional
+from typing import Optional
 
 from fastapi import APIRouter, Depends, HTTPException, Header
 from fastapi.responses import JSONResponse
-from pydantic import BaseModel, Field, field_validator
 
 from dependencies import get_current_user, get_authorized_agent
-from models import User
+from models import FanOutRequest, FanOutResponse, FanOutTaskResponse, User
 from services.fan_out_service import (
     FanOutService,
     FanOutTaskInput,
@@ -27,105 +25,6 @@ from services.platform_audit_service import platform_audit_service, AuditEventTy
 logger = logging.getLogger(__name__)
 
 router = APIRouter(prefix="/api/agents", tags=["fan-out"])
-
-# ---------------------------------------------------------------------------
-# Request / Response models
-# ---------------------------------------------------------------------------
-
-TASK_ID_RE = re.compile(r"^[a-zA-Z0-9_-]{1,64}$")
-MAX_TASKS = 50
-MAX_CONCURRENCY = 10
-
-
-class FanOutTask(BaseModel):
-    """A single task in a fan-out request."""
-    id: str
-    message: str = Field(..., min_length=1, max_length=100_000)
-
-    @field_validator("id")
-    @classmethod
-    def validate_task_id(cls, v: str) -> str:
-        if not TASK_ID_RE.match(v):
-            raise ValueError(
-                f"Task ID must be 1-64 alphanumeric characters, hyphens, or underscores: '{v}'"
-            )
-        return v
-
-
-class FanOutRequest(BaseModel):
-    """Request model for fan-out parallel task execution."""
-    tasks: List[FanOutTask]
-    agent: str = "self"
-    # Optional overall fan-out deadline. When None, no outer deadline is
-    # applied — each sub-task is still bounded by the target agent's
-    # configured execution_timeout_seconds (TIMEOUT-001).
-    timeout_seconds: Optional[int] = None
-    max_concurrency: int = 3
-    policy: str = "best-effort"
-    model: Optional[str] = None
-    system_prompt: Optional[str] = None
-    allowed_tools: Optional[List[str]] = None
-
-    @field_validator("tasks")
-    @classmethod
-    def validate_tasks(cls, v: List[FanOutTask]) -> List[FanOutTask]:
-        if len(v) == 0:
-            raise ValueError("At least one task is required")
-        if len(v) > MAX_TASKS:
-            raise ValueError(f"Maximum {MAX_TASKS} tasks per fan-out")
-        # Check for duplicate IDs
-        ids = [t.id for t in v]
-        if len(ids) != len(set(ids)):
-            dupes = [i for i in ids if ids.count(i) > 1]
-            raise ValueError(f"Duplicate task IDs: {set(dupes)}")
-        return v
-
-    @field_validator("max_concurrency")
-    @classmethod
-    def validate_concurrency(cls, v: int) -> int:
-        if v < 1 or v > MAX_CONCURRENCY:
-            raise ValueError(f"max_concurrency must be between 1 and {MAX_CONCURRENCY}")
-        return v
-
-    @field_validator("timeout_seconds")
-    @classmethod
-    def validate_timeout(cls, v: Optional[int]) -> Optional[int]:
-        if v is None:
-            return v
-        if v < 10 or v > 3600:
-            raise ValueError("timeout_seconds must be between 10 and 3600")
-        return v
-
-    @field_validator("policy")
-    @classmethod
-    def validate_policy(cls, v: str) -> str:
-        if v != "best-effort":
-            raise ValueError("Only 'best-effort' policy is supported")
-        return v
-
-
-class FanOutTaskResponse(BaseModel):
-    """Result of a single fan-out subtask."""
-    id: str
-    status: str
-    response: Optional[str] = None
-    error: Optional[str] = None
-    error_code: Optional[str] = None
-    execution_id: Optional[str] = None
-    cost: Optional[float] = None
-    context_used: Optional[int] = None
-    duration_ms: Optional[int] = None
-
-
-class FanOutResponse(BaseModel):
-    """Aggregated fan-out result."""
-    fan_out_id: str
-    status: str
-    total: int
-    completed: int
-    failed: int
-    results: List[FanOutTaskResponse]
-
 
 # ---------------------------------------------------------------------------
 # Endpoint
