+++ src/backend/routers/internal.py
@@ -12,12 +12,21 @@ import asyncio
 import os
 import hmac
 from fastapi import APIRouter, HTTPException, Depends, Request, Header
-from pydantic import BaseModel
-from typing import Optional, Dict, List
+from typing import Optional, Dict
 import logging
 
 from database import db
-from models import ActivityState, ActivityType, ShareFileRequest, ShareFileResponse, TaskExecutionStatus
+from models import (
+    ActivityCompleteRequest,
+    ActivityTrackRequest,
+    ActivityType,
+    InternalAuditRequest,
+    InternalTaskExecutionRequest,
+    ShareFileRequest,
+    ShareFileResponse,
+    TaskExecutionStatus,
+    ValidateExecutionRequest,
+)
 from services.activity_service import activity_service
 from services.task_execution_service import get_task_execution_service
 from services.platform_audit_service import platform_audit_service, AuditEventType
@@ -109,22 +118,6 @@ async def internal_agent_sync_health(agent_name: str):
 # =============================================================================
 # Activity Tracking Endpoints (for dedicated scheduler)
 # =============================================================================
-
-class ActivityTrackRequest(BaseModel):
-    """Request model for tracking activity start."""
-    agent_name: str
-    activity_type: str  # e.g., "schedule_start"
-    user_id: Optional[int] = None
-    triggered_by: str = "schedule"  # schedule, manual, user, agent, system
-    related_execution_id: Optional[str] = None
-    details: Optional[Dict] = None
-
-
-class ActivityCompleteRequest(BaseModel):
-    """Request model for completing an activity."""
-    status: str = ActivityState.COMPLETED  # ActivityState: completed, failed
-    details: Optional[Dict] = None
-    error: Optional[str] = None
 
 
 @router.post("/activities/track")
@@ -219,22 +212,6 @@ async def complete_activity(activity_id: str, request: ActivityCompleteRequest):
 # =============================================================================
 # Task Execution Endpoint (for dedicated scheduler)
 # =============================================================================
-
-class InternalTaskExecutionRequest(BaseModel):
-    """Request model for internal task execution via TaskExecutionService."""
-    agent_name: str
-    message: str
-    triggered_by: str = "schedule"
-    model: Optional[str] = None
-    timeout_seconds: Optional[int] = None  # TIMEOUT-001: None = use agent's config (default 15 min)
-    allowed_tools: Optional[List[str]] = None
-    execution_id: Optional[str] = None
-    async_mode: bool = False
-    # #171: optional schedule metadata surfaced in the agent's execution context block.
-    schedule_name: Optional[str] = None
-    schedule_cron: Optional[str] = None
-    schedule_next_run: Optional[str] = None
-    attempt: Optional[int] = None
 
 
 def _schedule_context_from(request: "InternalTaskExecutionRequest") -> Optional[Dict]:
@@ -455,16 +432,6 @@ async def _execute_task_internal_background(task_service, request: InternalTaskE
 # Validation Endpoints (VALIDATE-001)
 # =============================================================================
 
-class ValidateExecutionRequest(BaseModel):
-    """Request model for triggering execution validation."""
-    execution_id: str
-    agent_name: str
-    schedule_id: str
-    original_message: str
-    execution_response: str
-    custom_prompt: Optional[str] = None
-    timeout_seconds: int = 120
-
 
 @router.post("/validate-execution")
 async def validate_execution(request: ValidateExecutionRequest):
@@ -541,22 +508,6 @@ async def _run_validation_background(
 # =============================================================================
 # Audit Logging Endpoint (SEC-001 Phase 3 — MCP server integration)
 # =============================================================================
-
-class InternalAuditRequest(BaseModel):
-    """Request model for audit log entries from MCP server."""
-    event_type: str          # AuditEventType value
-    event_action: str        # e.g. "tool_call"
-    source: str = "mcp"      # Always "mcp" for MCP server calls
-    # MCP auth context
-    mcp_key_id: Optional[str] = None
-    mcp_key_name: Optional[str] = None
-    mcp_scope: Optional[str] = None
-    actor_agent_name: Optional[str] = None
-    # Target
-    target_type: Optional[str] = None
-    target_id: Optional[str] = None
-    # Details
-    details: Optional[Dict] = None
 
 
 @router.post("/audit")
